adds utility functions, reduces replacers code
